v3/services/chest-svc/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
onnxruntime==1.20.1
numpy==1.26.4
Pillow==11.1.0
boto3==1.35.86
httpx==0.28.1
scipy>=1.11.0
